Make vertical_position easier to read in the debugger. No reason for storage to be uint8.

+++ src/mpeg1_vid/mpeg1.hpp
@@ -58,7 +58,7 @@ namespace mpeg1 {
     };
 
     struct SliceHeader {
-        uint8_t vertical_position;
+        uint32_t vertical_position;
         uint8_t quantizer_scale;
     };
